properly check if user has no favorited services, break up components
