+++ web/src/app/alerts/AlertsList.js
@@ -1,4 +1,4 @@
-import React, { useState } from 'react'
+import React from 'react'
 import { PropTypes as p } from 'prop-types'
 import {
   Checkbox,
@@ -6,27 +6,22 @@ import {
   Hidden,
   Typography,
   makeStyles,
-  isWidthDown,
 } from '@material-ui/core'
 import { useDispatch, useSelector } from 'react-redux'
 import { absURLSelector, urlParamSelector } from '../selectors'
 import QueryList from '../lists/QueryList'
 import gql from 'graphql-tag'
 
-import CreateAlertFab from './CreateAlertFab'
 import AlertsListFilter from './components/AlertsListFilter'
 import AlertsListControls from './components/AlertsListControls'
 import AlertsCheckboxControls from './AlertsCheckboxControls'
+import AlertsListFloatingItems from './AlertsListFloatingItems'
 import statusStyles from '../util/statusStyles'
 import {
   setCheckedAlerts as _setCheckedAlerts,
   setAlerts as _setAlerts,
 } from '../actions'
 import { formatTimeSince } from '../util/timeFormat'
-import SnackbarContent from '@material-ui/core/SnackbarContent'
-import InfoIcon from '@material-ui/icons/Info'
-import Snackbar from '@material-ui/core/Snackbar'
-import useWidth from '../util/useWidth'
 
 export const alertsListQuery = gql`
   query alertsList($input: AlertSearchOptions) {
@@ -53,20 +48,6 @@ export const alertsListQuery = gql`
 `
 
 const useStyles = makeStyles(theme => ({
-  snackbar: {
-    backgroundColor: theme.palette.primary['500'],
-    height: '6.75em',
-    width: '20em', // only triggers on desktop, 100% on mobile devices
-  },
-  snackbarIcon: {
-    fontSize: 20,
-    opacity: 0.9,
-    marginRight: theme.spacing(1),
-  },
-  snackbarMessage: {
-    display: 'flex',
-    alignItems: 'center',
-  },
   checkbox: {
     marginRight: 'auto',
   },
@@ -91,30 +72,20 @@ function getStatusFilter(s) {
 
 export default function AlertsList(props) {
   const classes = useStyles()
-  const width = useWidth()
-  const isFullScreen = isWidthDown('md', width)
-
-  // always open unless clicked away from or there are services present
-  const [snackbarOpen, setSnackbarOpen] = useState(true)
 
   // get redux vars
   const absURL = useSelector(absURLSelector)
   const params = useSelector(urlParamSelector)
-  const actionComplete = useSelector(state => state.alerts.actionComplete)
   const allServices = params('allServices')
   const checkedAlerts = useSelector(state => state.alerts.checkedAlerts)
   const filter = params('filter', 'active')
-  const isFirstLogin = params('isFirstLogin')
 
   // setup redux actions
   const dispatch = useDispatch()
   const setCheckedAlerts = arr => dispatch(_setCheckedAlerts(arr))
   const setAlerts = arr => dispatch(_setAlerts(arr))
 
-  // todo: need noFavorites?
-  const showFavoritesWarning =
-    snackbarOpen && !allServices && !props.serviceID && !isFirstLogin
-
+  // alerts list query variables
   const variables = {
     input: {
       filterByStatus: getStatusFilter(filter),
@@ -153,12 +124,6 @@ export default function AlertsList(props) {
         return classes.statusError
       default:
         return classes.noStatus
-    }
-  }
-
-  function handleCloseSnackbar(event, reason) {
-    if (reason === 'clickaway') {
-      setSnackbarOpen(false)
     }
   }
 
@@ -214,32 +179,7 @@ export default function AlertsList(props) {
           </Hidden>
         }
       />
-      <Snackbar
-        anchorOrigin={{
-          vertical: 'bottom',
-          horizontal: 'left',
-        }}
-        open={showFavoritesWarning}
-        onClose={handleCloseSnackbar}
-      >
-        <SnackbarContent
-          className={classes.snackbar}
-          aria-describedby='client-snackbar'
-          message={
-            <span id='client-snackbar' className={classes.snackbarMessage}>
-              <InfoIcon className={classes.snackbarIcon} />
-              It looks like you have no favorited services. Visit your most used
-              services to set them as a favorite, or enable the filter to view
-              alerts for all services.
-            </span>
-          }
-        />
-      </Snackbar>
-      <CreateAlertFab
-        serviceID={props.serviceID}
-        showFavoritesWarning={showFavoritesWarning}
-        transition={isFullScreen && (showFavoritesWarning || actionComplete)}
-      />
+      <AlertsListFloatingItems serviceID={props.serviceID} />
     </React.Fragment>
   )
 }
